Moved DB logic to separate package
